resolve bug add comments
