+++ target_chains/cosmwasm/tools/package.json
@@ -24,15 +24,14 @@
     "chain-registry": "^1.6.0",
     "cosmjs-utils": "^0.1.0",
     "dotenv": "^16.0.0",
-    "download": "^8.0.0",
     "ethers": "^5.4.4",
+    "node-downloader-helper": "^2.1.7",
     "osmojs": "^13.0.0-rc.7-i-alpha",
     "rimraf": "^5.0.0",
     "yargs": "^17.0.1"
   },
   "devDependencies": {
     "@types/adm-zip": "^0.5.0",
-    "@types/download": "^8.0.2",
     "@types/yargs": "^17.0.18",
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "@typescript-eslint/parser": "^5.43.0",
